# packages/schemas/openapi.v1.yaml
openapi: 3.1.0
info:
  title: MARID Control Plane API
  version: 0.3.0
  description: Tenant-scoped control plane. Execution records do not execute tools.
servers:
  - url: http://localhost:5056
paths:
...
  /v1/retention/policies:
    get:
      summary: Read tenant retention periods by category
      security: [{ bearerAuth: [] }]
      x-required-scope: retention.read
      responses:
        '200': { description: Retention policies }
  /v1/retention/policies/{category}:
    put:
      summary: Change a tenant retention period
      security: [{ bearerAuth: [] }]
      x-required-scope: retention.admin
      parameters:
        - { name: category, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Policy changed and audited }
        '403': { description: Requires a human incident manager }
  /v1/evidence/{evidenceId}/holds:
    post:
      summary: Apply a tenant legal hold to live raw evidence
      security: [{ bearerAuth: [] }]
      x-required-scope: retention.admin
      parameters:
        - { name: evidenceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Legal hold applied or already active }
        '404': { description: No live evidence in the token tenant }
  /v1/evidence/{evidenceId}/holds/{holdId}/release:
    post:
      summary: Release a tenant legal hold with an audit event
      security: [{ bearerAuth: [] }]
      x-required-scope: retention.admin
      parameters:
        - { name: evidenceId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: holdId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Hold released }
        '404': { description: No active hold in the token tenant }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT